fixed bug

+++ hyd/HaukursYouTubeDownloader.py
@@ -5,12 +5,12 @@ from time import time
 import os
 
 def main():
+    global mp3IsClicked
+    mp3IsClicked = False
+
     window = Tk()
     path = os.path.join(os.path.join(os.environ["HOME"]), 'YoutubeDownloads') # Make path
     os.makedirs(path, exist_ok=True) # Create folder if it doesn't exist already
-    start_time = 0 
-    end_time = 0
-    mp3IsClicked = False
 
     # The window itself
     window.geometry('600x150')
@@ -77,13 +77,12 @@ def main():
                 fixedtitle = fixedtitle.replace(char, "")
             start_time = time()
             YouTube(link).streams.get_highest_resolution().download(path)
-            if os.path.exists(f"{path}/{fixedtitle}.mp4"):
-                if mp3IsClicked == True:
-                    video = VideoFileClip(f"{path}/{fixedtitle}.mp4")
-                    os.remove(f"{path}/{fixedtitle}.mp4")
-                    video.audio.write_audiofile(f"{path}/{fixedtitle}.mp3")
-                    mp3IsClicked = False
-            end_time = time()
+            if os.path.exists(f"{path}/{fixedtitle}.mp4") and mp3IsClicked:
+                video = VideoFileClip(f"{path}/{fixedtitle}.mp4")
+                os.remove(f"{path}/{fixedtitle}.mp4")
+                video.audio.write_audiofile(f"{path}/{fixedtitle}.mp3")
+                mp3IsClicked = False     
+            end_time = time()           
             popup = Tk()
             popup.title("Download Success!")
             popup.geometry("350x125")
@@ -101,6 +100,10 @@ def main():
             errorLabel = Label(error, text="Download Failed :/ \n Please check your internet connection.",  fg="#EDEDED", bg="#222222")
             errorLabel.place(relx=0.5, rely=0.25, anchor=CENTER)
             error.mainloop()
+            
+
+
+        
 
     window.mainloop()
 if __name__ == '__main__':
